+++ src/pages/post/PostAnswerPage.jsx
@@ -1,5 +1,6 @@
 import { useParams, useRouteLoaderData } from "react-router-dom";
-import useQuestions from "./components/useQuestions";
+import useQuestions from "./hooks/useQuestions";
+import useSubject from "./hooks/useSubject";
 import {
   FeedDeleteButton,
   FeedList,
@@ -8,7 +9,6 @@ import {
   FeedCard,
   Message,
 } from "@components/FeedCard";
-import useSubject from "./components/useSubject";
 import { Notify } from "@components/Toast";
 import { MESSAGES } from "@constants/messages";
